+++ test/libsolidity/smtCheckerTests/imports/imported_fail_1.sol
@@ -1,4 +1,4 @@
-==== Source:  ====
+==== Source: C.sol ====
 import "B.sol";
 pragma experimental SMTChecker;
 contract C is B {
@@ -21,5 +21,4 @@ contract B is A {
 	}
 }
 // ----
-// Warning 6328: (103-117): CHC: Assertion violation happens here.
-// Warning 6328: (B.sol:71-85): CHC: Assertion violation happens here.
+// Warning 6328: (C.sol:103-117): CHC: Assertion violation happens here.